+++ src/arch/x86/amd64_syscall.c
@@ -3,11 +3,11 @@
 #include "../../../include/min/min_def.h"
 // rax rdi rsi rdx r10 r8 r9
 
-ssize_t sys_read(i32 fd, void *buf, u64 size) {
+ssize_t sys_read(i32 fd, void *buf, size_t size) {
 	register u64         rax __asm__("rax") = NR_READ;
 	register i32         rdi __asm__("rdi") = fd;
 	register const void *rsi __asm__("rsi") = buf;
-	register u64         rdx __asm__("rdx") = size;
+	register size_t      rdx __asm__("rdx") = size;
 	ssize_t ret;
 	__asm__ volatile ( "syscall"
 		: "=a" (ret)
@@ -17,11 +17,11 @@ ssize_t sys_read(i32 fd, void *buf, u64 size) {
 	return ret;
 }
 
-ssize_t sys_write(i32 fd, const void *buf, u64 size) {
+ssize_t sys_write(i32 fd, const void *buf, size_t size) {
 	register u64         rax __asm__("rax") = NR_WRITE;
 	register i32         rdi __asm__("rdi") = fd;
 	register const void *rsi __asm__("rsi") = buf;
-	register u64         rdx __asm__("rdx") = size;
+	register size_t      rdx __asm__("rdx") = size;
 	ssize_t ret;
 	__asm__ volatile ( "syscall"
 		: "=a" (ret)
@@ -31,15 +31,15 @@ ssize_t sys_write(i32 fd, const void *buf, u64 size) {
 	return ret;
 }
 
-void *sys_mmap(void *addr, u32 length, i32 prot,
+void *sys_mmap(void *addr, size_t length, i32 prot,
                i32 flags, i32 fd, i32 off) {
-	register u64   rax __asm__("rax") = NR_MMAP;
-	register void *rdi __asm__("rdi") = addr;
-	register u32   rsi __asm__("rsi") = length;
-	register i32   rdx __asm__("rdx") = prot;
-	register i32   r10 __asm__("r10") = flags;
-	register i32   r8  __asm__("r8") = fd;
-	register i32   r9  __asm__("r9") = off;
+	register u64     rax __asm__("rax") = NR_MMAP;
+	register void   *rdi __asm__("rdi") = addr;
+	register size_t  rsi __asm__("rsi") = length;
+	register i32     rdx __asm__("rdx") = prot;
+	register i32     r10 __asm__("r10") = flags;
+	register i32     r8  __asm__("r8") = fd;
+	register i32     r9  __asm__("r9") = off;
 	void *ret;
 	__asm__ volatile ( "syscall"
 		: "=a" (ret)
@@ -49,10 +49,10 @@ void *sys_mmap(void *addr, u32 length, i32 prot,
 	return ret;
 }
 
-ssize_t sys_munmap(void *addr, u32 length) {
-	register u64   rax __asm__("rax") = NR_MUNMAP;
-	register void *rdi __asm__("rdi") = addr;
-	register u32   rsi __asm__("rsi") = length;
+ssize_t sys_munmap(void *addr, size_t length) {
+	register u64     rax __asm__("rax") = NR_MUNMAP;
+	register void   *rdi __asm__("rdi") = addr;
+	register size_t  rsi __asm__("rsi") = length;
 	ssize_t ret;
 	__asm__ volatile ( "syscall"
 		: "=a" (ret)
@@ -76,11 +76,11 @@ ssize_t sys_socket(i32 domain, i32 type, i32 protocol) {
 	return ret;
 }
 
-ssize_t sys_bind(i32 sockfd, sockaddr *my_addr, i32 addrlen) {
+ssize_t sys_bind(i32 sockfd, sockaddr *my_addr, ssize_t addrlen) {
 	register u64       rax __asm__("rax") = NR_SOCKET;
 	register i32       rdi __asm__("rdi") = sockfd;
 	register sockaddr *rsi __asm__("rsi") = my_addr;
-	register i32       rdx __asm__("rdx") = addrlen;
+	register ssize_t   rdx __asm__("rdx") = addrlen;
 	ssize_t ret;
 	__asm__ volatile ( "syscall"
 		: "=a" (ret)
@@ -125,11 +125,11 @@ clock_t sys_times(tms *t) {
 	return ret;
 }
 
-ssize_t sys_getrandom(char *buf, u64 count, u32 flags) {
-	register u64   rax __asm__("rax") = NR_GETRANDOM;
-	register char *rdi __asm__("rdi") = buf;
-	register u64   rsi __asm__("rsi") = count;
-	register u32   rdx __asm__("rdx") = flags;
+ssize_t sys_getrandom(char *buf, size_t count, u32 flags) {
+	register u64     rax __asm__("rax") = NR_GETRANDOM;
+	register char   *rdi __asm__("rdi") = buf;
+	register size_t  rsi __asm__("rsi") = count;
+	register u32     rdx __asm__("rdx") = flags;
 	ssize_t ret;
 	__asm__ volatile ( "syscall"
 		: "=a" (ret)
